add test auth in method post and add new file
