+++ tests/java/test_api_rest.py
@@ -14,112 +14,6 @@ exampleTest = 'tests/java/example_java/PruebaTest.java'
 def delete_db():
 	db.session.query(Challenge_java).delete()
 
-# insert a valid challenge and return a status code equal to 200
-def test_post_java(client):
-	delete_db()
-	location = '/java/java-challenges'
-	url = 'http://localhost:5000/java/java-challenges'
-
-	data = createChallenge('example-challenges/java-challenges/Median.java','example-challenges/java-challenges/MedianTest.java','Median','MedianTest', 'pass', '1')
-	
-	token=get_token(client)
-	#resp = client.post(url, data=data)
-	resp1=client.post(url, headers={'Authorization': f'JWT {token}'}, data=data)
-	#assert resp.status_code == 401
-	assert resp1.status_code==200
-	
-# insert a valid challenge and get challenge
-def test_post_get_all(client):
-	delete_db()
-	url = 'http://localhost:5000/java/java-challenges'
-
-	data = createChallenge('example-challenges/java-challenges/Median.java','example-challenges/java-challenges/MedianTest.java','Median','MedianTest', 'pass', '1')
-	client.post(url, data=data)
-	resp = client.get(url)
-	response = json.loads(json.dumps(resp.json))
-	a = response['challenges'][0]
-	
-	assert resp.status_code == 200
-	assert len(response['challenges']) == 1
-	assert a['repair_objective'] == 'pass'
-	assert a['complexity'] == 1
-	assert a['id'] != 0
-
-# insert two valid challenges and then get them and compare the results
-def test_many_loads(client):
-	delete_db()
-	url = url = 'http://localhost:5000/java/java-challenges'
-
-	data = createChallenge('example-challenges/java-challenges/Median.java','example-challenges/java-challenges/MedianTest.java','Median','MedianTest', 'pass', '1')
-	data2 = createChallenge('tests/java/example_java/Prueba.java','tests/java/example_java/PruebaTest.java','Prueba','PruebaTest','Pasa','3')
-	
-	client.post(url, data=data)
-	p = client.post(url, data=data2)
-
-	resp = client.get(url)
-	response = json.loads(json.dumps(resp.json))
-	b = response['challenges']
-	
-	assert p.status_code == 200
-	assert resp.status_code == 200
-	assert len(response['challenges']) == 2
-	assert b[0]['repair_objective'] == 'pass'
-	assert b[1]['repair_objective'] == 'Pasa'
-
-# insert two valid challenges equal, first status code equal 200 and second equal 404
-def test_upload_exist(client):
-	delete_db()
-	url = url = 'http://localhost:5000/java/java-challenges'
-	data = createChallenge('example-challenges/java-challenges/Median.java','example-challenges/java-challenges/MedianTest.java','Median','MedianTest', 'pass', '1')
-	datab = createChallenge('example-challenges/java-challenges/Median.java','example-challenges/java-challenges/MedianTest.java','Median','MedianTest', 'pass', '1')
-	resp = client.post(url, data=data)
-	resp2 = client.post(url, data=datab)
-
-	assert resp.status_code == 200
-	assert resp2.status_code == 404
-
-# insert java class that does not compile
-def test_file_not_compile_class(client):
-	delete_db()
-	url = url = 'http://localhost:5000/java/java-challenges'
-	data = createChallenge('tests/java/example_java/Nocompile.java','tests/java/example_java/NocompileTest.java', 'Nocompile', 'NocompileTest', 'nada', '2')
-	try:
-		resp = client.post(url, data=data)
-	except Exception as e:
-		assert str(e) == "Algun archivo no compila o pasa todos los test, debe fallar algun test para cargar"
-	assert resp.status_code == 404
-	
-# insert java test suite that does not compile
-def test_file_not_compile_test(client):
-	delete_db()
-	url = url = 'http://localhost:5000/java/java-challenges'
-	data = createChallenge('tests/java/example_java/Testfailclass.java','tests/java/example_java/Testfailtest.java', 'Testfailclass', 'Testfailtest', 'cosa', '3')
-	try:
-		resp = client.post(url, data=data)
-	except Exception as e:
-		assert str(e) == "Algun archivo no compila o pasa todos los test, debe fallar algun test para cargar"		
-	assert resp.status_code == 404
-
-# insert valid challenge that pass all test
-def test_pass_all_test(client):
-	delete_db()
-	url = url = 'http://localhost:5000/java/java-challenges'
-	data = createChallenge('tests/java/example_java/Passalltest.java','tests/java/example_java/Passalltesttest.java', 'Passalltest', 'Passalltesttest', 'dale', '5')
-	try:
-		resp = client.post(url, data=data)
-	except Exception as e:
-		assert str(e) == "Algun archivo no compila o pasa todos los test, debe fallar algun test para cargar"
-	assert resp.status_code == 404
-
-def test_get_java(client):
-	delete_db()
-	url = 'http://localhost:5000/java/java-challenges'
-	resp = client.get(url)
-	a = resp.json
-	
-	assert resp.status_code == 200
-	assert a['challenges'] == []
-	
 def test_get_Id_after_post(client):
 	db.session.query(Challenge_java).delete()
 	url = 'http://localhost:5000/java/java-challenges'
